+++ src/common/utils/index.js
@@ -1,5 +1,4 @@
 export * from './createClearFilterHandler';
-export * from './createOnChangeSelectionFilter';
 export * from './getActiveFilters';
 export * from './getFundOptions';
 export * from './getMaterialTypesOptions';